# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-timeframe signal API (daily direction + intraday timing)",
        "title": "Midas Touch API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/candles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "market"
                ],
                "summary": "Get candles",
                "parameters": [
                    {
                        "type": "string",
                        "default": "NVDA",
                        "description": "Ticker symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "1d",
                        "description": "1d,5,15,30,60,120,240",
                        "name": "timeframe",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 300,
                        "description": "rows",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mongodb.CandleDoc"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/notify": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signal"
                ],
                "summary": "Evaluate and send Telegram signal",
                "parameters": [
                    {
                        "type": "string",
                        "default": "NVDA",
                        "description": "Ticker symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "120",
                        "description": "60,120,240",
                        "name": "timing_tf",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/signal": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signal"
                ],
                "summary": "Evaluate a signal",
                "parameters": [
                    {
                        "type": "string",
                        "default": "NVDA",
                        "description": "Ticker symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "120",
                        "description": "60,120,240",
                        "name": "timing_tf",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sources/status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "External data source status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.SourceStatus"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.SourceStatus": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "latency_ms": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "mongodb.CandleDoc": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "timeframe": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "volume": {
                    "type": "number"
                }
            }
        }
    }
}